+++ test/tests/column_test.dart
@@ -6,24 +6,21 @@ void main() {
   group('DaviColumn', () {
     test('setting sort', () {
       DaviColumn column = DaviColumn(id: 'id', sortable: true);
-      DaviSort sort = DaviSort('id');
-      DaviColumnHelper.setSort(column: column, sort: sort, priority: 2);
+      DaviColumnHelper.setSort(
+          column: column, direction: DaviSortDirection.ascending, priority: 2);
       expect(column.sortPriority, 2);
-      sort = DaviSort('x');
-      expect(
-          () =>
-              DaviColumnHelper.setSort(column: column, sort: sort, priority: 3),
-          throwsA(const TypeMatcher<ArgumentError>()));
 
       column = DaviColumn(id: 'id', sortable: false);
-      sort = DaviSort('id');
       expect(
-          () =>
-              DaviColumnHelper.setSort(column: column, sort: sort, priority: 1),
+          () => DaviColumnHelper.setSort(
+              column: column,
+              direction: DaviSortDirection.ascending,
+              priority: 1),
           throwsA(const TypeMatcher<ArgumentError>()));
 
       column = DaviColumn(id: 1);
-      expect(DaviColumnHelper.setSortPriority(column: column, value: 1), false);
+      expect(
+          DaviColumnHelper.setSortPriority(column: column, priority: 1), false);
     });
   });
 }